requirements.txt
openpyxl>=3.1.0
pandas>=2.0.0
langchain>=0.1.0
langchain-anthropic>=0.1.0
chromadb>=0.4.0
sentence-transformers>=2.0.0
python-dotenv>=1.0.0
streamlit>=1.30.0
watchdog>=3.0.0
torchvision>=0.26.0

# Multi-Provider LLM Support
anthropic>=0.40.0
google-generativeai>=0.8.0
openai>=1.0.0
